Feat and Refactor: Update Metric Feat and remove unused metric update logic
This commit implements update metric by utilizing updateExerciseService to update the Exercise's since metric is a json data in the DB.
This commit also removes redundant functions related to updating
metrics in the `Exercise` component.  The update logic is handled directly within the `ExerciseSet` component, simplifying
the code and improving maintainability.  No breaking changes
are introduced as the functionality remains the same.

Signed-off-by: Jericho Pasco <112312834+Jpasco31@users.noreply.github.com>