fix: Improve exercise position calculation
This commit improves the calculation of the position of newly added exercises.  Previously, the position was simply the total number of exercises, leading to incorrect ordering, especially for nested exercises.  Now, the position is calculated based on the parent exercise and date, ensuring exercises are correctly ordered within their respective parent and date groups.

Signed-off-by: Jericho Pasco <112312834+Jpasco31@users.noreply.github.com>